ui/kanban/requirements.txt
click>=8.1.3
rich>=12.6.0
pyyaml>=6.0
pytest>=7.2.0
